requirements.txt
# Web Framework
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
email-validator==2.1.0.post1

# Data Processing
pandas==2.2.1
numpy==1.26.4
dask[complete]==2025.5.1
pyarrow==15.0.0

# API Documentation
pydantic==2.6.3
pydantic-settings==2.2.1

# Utilities
python-dotenv==1.0.1
PyYAML==6.0.1
ijson==3.2.3
tqdm==4.66.2
psutil==5.9.8

# Testing
pytest==8.0.2
pytest-asyncio==0.23.5
httpx==0.27.0

# Development
black==24.2.0
isort==5.13.2
flake8==7.0.0
mypy==1.8.0
